move fieldIsEnum to base-application/support
