+++ generators/base-application/support/index.mts
@@ -1,0 +1,1 @@
+export * from './field-utils.mjs';
