+++ generators/bootstrap-application/generator.mts
@@ -28,7 +28,7 @@ import {
   preparePostEntityCommonDerivedProperties,
   preparePostEntityServerDerivedProperties,
 } from '../../utils/entity.mjs';
-import { fieldIsEnum } from '../entity/support/index.mjs';
+import { fieldIsEnum } from '../base-application/support/index.mjs';
 import { GENERATOR_BOOTSTRAP_APPLICATION_CLIENT, GENERATOR_BOOTSTRAP_APPLICATION_SERVER } from '../generator-list.mjs';
 
 import type { ClientServerApplication } from '../common/types.mjs';
