+++ generators/entity/support/index.mjs
@@ -22,6 +22,5 @@ export {
   isSignedNumber as inputIsSignedNumber,
   isSignedDecimalNumber as inputIsSignedDecimalNumber,
 } from './asserts.mjs';
-export { fieldIsEnum } from './field-utils.mjs';
 export { getPkType } from './pk-utils.mjs';
 export { getEntityFolderName } from './template-utils.mjs';
